refactor: dedupe formatUptime between AdminPage and SuperuserPage (#1053)
Both pages had byte-identical copies of formatUptime(). Extracted to
lib/utils.ts and imported in both places. Fixes the 10-line duplicate
block SonarCloud flagged between these two files (one of the block
pairs surfaced when #1052's button-type sweep touched lines inside
it -- see #1053).

No behavior change: same function body, moved verbatim.
